requirements.txt
Flask
spinqit
python-dotenv
numpy==1.23.5
